Patch from Joachim Metz <joachim.metz@gmail.com> for 64 bit compile. Signed/unsigned cleanup from 'CFLAGS=-Wextra ./configure' Reindent vbuf.c to make it readable.

+++ src/debug.c
@@ -47,7 +47,7 @@ void pst_debug(char *fmt, ...) {
 
 #define NUM_COL 30
 void pst_debug_hexdumper(FILE *out, unsigned char *buf, size_t size, int col, int delta) {
-    int off = 0, toff;
+    size_t off = 0, toff;
     int count = 0;
 
     if (!out) return;   // no file
